feat: useReportComponent로 로직 분리
